+++ package.json
@@ -11,7 +11,7 @@
   "dependencies": {
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
-    "sass": "^1.62.0"
+    "sass": "^1.62.0",
     "react-router-dom": "^6.10.0"
   },
   "devDependencies": {
